pyproject.toml
[project]
name = "arxiv-mcp-improved"
version = "0.2.0"
description = "An improved async MCP server to fetch and process ArXiv papers with parallel downloads and TeX generation."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "diskcache>=5.6.3",
    "markitdown[pdf]>=0.1.2",
    "mcp[cli]>=1.9.3",
    "httpx[http2]>=0.27.0",
    "aiolimiter>=1.1.0",
    "tenacity>=8.2.3",
    "aiofiles>=23.2.0",
    "python-magic>=0.4.27",
    "pyyaml>=6.0",
    "aiohttp>=3.8.0",
    "python-dotenv>=1.1.1",
]

[project.scripts]
arxiv-mcp-improved = "arxiv_mcp:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/arxiv_mcp"]
